+++ src/qt/transactiontablemodel.h
@@ -1,8 +1,6 @@
 #ifndef TRANSACTIONTABLEMODEL_H
 #define TRANSACTIONTABLEMODEL_H
 
-#include "darksilkunits.h"
-
 #include <QAbstractTableModel>
 #include <QStringList>
 
@@ -22,11 +20,11 @@ public:
 
     enum ColumnIndex {
         Status = 0,
-        Watchonly = 1,
-        Date = 2,
-        Type = 3,
-        ToAddress = 4,
-        Amount = 5
+        Date = 1,
+        Type = 2,
+        ToAddress = 3,
+        Amount = 4,
+        Watchonly = 5
     };
 
     /** Roles to get specific information from a transaction row.
